updated project structure and display image after taking photo
